refactor(仪表板): 外部参数设置逻辑优化
